completely replace the console
